+++ nbproject/project.properties
@@ -73,7 +73,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=Proyecto_Aseguradora
-main.class=FrontEnd.SucursalesVistaAdmin
+main.class=FrontEnd.VistaLogin
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
